MTS29: Updated PUT method for movie to handle theatre updates
