+++ src/main/resources/application.properties
@@ -5,7 +5,7 @@ spring.datasource.username=root
 spring.datasource.password=
 spring.datasource.driver-class-name=com.mysql.cj.jdbc.Driver
 spring.jpa.properties.hibernate.dialect=org.hibernate.dialect.MySQL8Dialect
-server.port=8080
+server.port=8081
 cors.originUrl=http://localhost:4200
 
 # Add these properties for debugging sql operations
